refactor(app-flow): Renames (2/2) (promote [cap-server]) ((1757))
Discussion: it was awkward having an "html" node under storage adapters
whose only purpose was to do form generation. "app-flow" has become a
useful catch-all for this stuff.

Once we freed app-flow out from under kiss it became both forbidden and
weird to have cap-server under kiss.

This is all a precursor to what we really want to get to: cleaning
up cap-server and abstracting something out of it.

(times)
   08:03  (from Hyperion) (minus 10 minutes) begin move files
   10:02  done!