refactor(app-flow): Rename (1 of 2)
Reason: When we started "app-flow", the question was, "what if your application
was generated from an app (stored in a database)"? Since then we have somewhat
loosened this lofty and perhaps misconceived goal to be more, "let's
represent the app-graph in code (somehow) at first, and also do all the other
clever mechanics first (get mounted by flask, maintain sane CLI-ability,
maintain being mounted by redbean)"

Also, in its infancy app-flow was just for generating a visualization of
the application's *design*, not so much for supporting the actual build-out
of the application.

Anyway, this file rename reflects that change in focus: tools for building
a graph (GraphViz) from a database do not need to be in the main center module
of this module.

Additionally:

+ README.md gets promoted to be in own doc/ directory, anticipating
  the coming grown of app-flow)
+ Document nodes that used to live in [kss] now live in [app-flow] to give
  [kss] more space (but note we are piggy-backing [cap-server] docs on to
  [app-flow])