abstract(cap-server->app-flow): (1 of 2): Abstract reusable logic ((1757))
We're breaking this into two commits to preserve interesting history.
In this first commit, we simply re-house the code (although it had some
significant architecture change here too: the "parent node" thing is now
a mandatory assumption; now there's this "derived values" structure).

In the next commit we'll refactor the same code so it's not as nasty
to read (finally!)

(times)
  11:44 (from Gaylord/Waters MI) begin "forming" module
  19:59 now