snapshot(tilex): Refactor into first-ever `fetch` call to backend
Justification: This commit doesn't add anything interesting *except*
it's a proof-of-concept of how we can use the new-ish `fetch` method
to do a backend request. Next we need to actually make the backend.

We are commiting this as-is so we can have a simple, barely-working
demonstration of this backend "call" when we need to troubleshoot
such a thing across browsers.

(times)
  08-07 20:15  (from SWP) begin learning about fetch
  08-08 18:48  (from SWP) working out details with fetch
        19:44  static API call to backend is good