script(STM): Create script for cleaning up flutter-generated filetree
I keep thinking about the scene from Lost where Dominic Monaghan's
character is stuck in a dark cave as a metaphor for his addiction
and he finds the way out by following a moth that's flying towards
a narrow slit of light. This is what it has been like making all of
the compromises to get to the point reflected in this commit and the
next.

(times)
  02-16 20:05  begin barebones app step
        20:13  status: codelab is big and gonna take time
  02-17 08:35  (from Hyperion) maybe start codelab. (nope, work)
  02-18 07:21  (ok for real this time) (from Hyperion) begin codelab
        07:50  (log) can't be "STM" must be "stm". Abandoned web for now.
               Experimenting with making subtree. But actually take break.
  02-19 08:22  (from H) add to work network chart
  02-20 08:14  (from H) lets actually make that script
        08:52  suspend for now to go do day-off things
        11:14  (from Stray Hen) resume
        11:58  (suspend to move to Cahoots)
        13:05  (resume from Cahoots)
        14:04  omg worked (linux desktop, with splitter script)
        14:31  cleaned up this worklog