visual(STM): Add vendor-generated files only
Discussion: We want a snapshot of the vendor-generated files exactly
as-generated except we don't want extraneous files (of which there are
many) which is why we made the script added in the previous commit.

This commit is a "visual test" because (at writing) ("works on my machine")
this commit of our flutter app ran as a linux desktop app on our linux
development box out-of-the-box as-is (alhamdulillah).

It's worth mentioning we did *not* yet try running the flutter automated
tests (although we did add the generated test file, because we intend
to use it).

It's also worth mentioning that the flutter "create" command wouldn't
let us name our app "STM" (can't start with a cap? can't be all-caps?)
but "stm" was valid. However, this conflicts with our own internal
name conventions ([#bs-029] (near mentions of "acronym")) which dictates
that we use all-caps on the filesystem IFF acronyms. Because our visual
test worked doing it our way, we are doing it our way for now.