test(STM): Add example test from cookbook (1)
This is an arbitrarily minimal test with dummy "asset" code
just for us to demonstrate to ourselves that we know how to
run our unit tests.

We brute force tried different version of the test package
backwards until we found one that didn't have dependency hell
issues. (I.e. we don't know what we are doing yet, in this
regard.)

Took 15 seconds to run. All test passed.

(times)
  16:10  (from A) begin unit test intro
  16:36  worked