step(STM): Change example table to 'Like' (& schema) (db, 3, v)
What do these new tags mean in this commit:

When a commit message is "tagged" with "db" (as this one is),
it means that this commit introduces a database schema change,
and so you should remove the previous database. (Yes this blows
away your data but there should be no data at this point.)
How-to below.

When a commit message is tagged with "v" it means
"visual test was passsed in this commit". It is undefined what
such a visual test entails. For us for this commit and ones near
it, it means we ran our app and clicked through some of the
interface (most importantly, testing the one operation that will
write to our database).

We'll explain what the number "3" here is below..

Running the tests:

We have changed table names from the previous commit and before,
so if you're coming from that direction (and your database exists
on the filesystem and is "stale" in its schema), remove the file

```bash
$ rm .dart_tool/sqflite_common_ffi/databases/app_database.db
```

Now, to run the tests:

```bash
$ flutter test test/replication_test.dart
```

There's a whole test file we are ignoring (next to the file above).

Three tests should have run and all three passed, hence why there's
"3" as a "tag" in this commit.