step(STM): Carefully change the name convention (3, v)
We're nervous about this one change because we don't yet
understand how much of things are baked in by the code generation,
hence we are isolating the change here in a dedicated commit.

We discuss how we do the casing in a name that has an acronym in it in
the ancient [#bs-029]. It is not yet set in stone.

- changed likeDao to like_DAO (the getter)
- changed like_DAO to likeDAO