step(STM): Set up second side of bridge (3, v)
We have been ruminating over this for over a week. For whatever personal
mental health reason, it felt literally unimaginable that we would be able
to "bridge the two sides": On the one hand, the side of the zen rock garden
of the demo code. On the other side is the scary, messy new world of async
and sqlite.

Probably the main reason we met this with such mental/emotional excitement
was that this would be the first time where we are really writing code,
out in the wild on our own, without training wheels.

We continue to procede with caution because still we feel like we have no
idea what we are doing.

In other words, we are squarely in the "impostor syndrome" phase of the journey.

About to be the man.

MAYBE PUT THIS INTO THE DOC and give it its own identifier

googled "dart call async from sync"

https://stackoverflow.com/questions/28238161/how-to-make-an-asynchronous-dart-call-synchronous

(times)
  04-24 12:04  (from Cahoots) clean up one final commit
        08:15  (from H) planning CRUD parallel (day idk)
  04-29 07:34  (from H) we can do it, use database to populate thing
        07:50  (ok wow our commit goal will work!?)