upgrade(STM): Attempt to upgrade all packages (3, v, SEE)
Synopsis:
+ In order to troubleshoot what we're attempting in the next step
  after this, we attempted to upgrade all our packages. We approached
  this by removing *all* explicit version numbers from pubspec.yaml
  where possible. (We call it "autobleed", the imaginary service that
  automatically upgrades packages and looks at what breaks. This is
  a step towards that.
+ (Certainly, if an under-development package goes into production,
  it should *not* be on autobleed. But we are a long way away from that.)
+ It seems that this upgrade advanced sqlparser by one minor version
  which seems to have broken its attempts at parsing our code that comes
  as an example from google. This commit has a fix for that (more below).
+ (SEE) Our generated schema shows a new problem: We want the
  auto-increment behavior when doing CREATE; we don't know how to
  get that without passing NULL as the id; we don't know how to do
  that unless we make that field nullable; WHICH in turn has impact
  in how the 'CREATE TABLE' is generated which in turn may create the
  column as not the special sqlite magic "integer not null primary key"
  etc. What we suspect is the case is we need the CREATE TABLE magic
  from *before* we introduced this ID field as nullable (a bad workaround.
  we need to find the correct fix for this.)

Narrative:

OK so here's a story: I got a runtime error from something
"floor" related, when attempting to implement 'DELETE' (which is itself
a story)..

+ Do I learn to step-debug into vendor code?
+ Where are the dart files of a package stored locally?
+ What if I need to step-debug into a vendor's package?

I decided to "back up a step" and attempt to get all packages upgraded.
This seems to have revealed a bit of a dependency hell I'm in; but I
don't yet know enough to assess this.

What was unimaginable was that a query so simple could be messing up
their SQL parser. My query was simply:

    Select * from Like

Now, *I* see something suspicious there; but I left the SQL in there
EXACTLY as the demo had it.

Anyway, what it seems like happened is the "sqlparser" package
correctly made itself more strict like I recommended. I CORRECTLY
GUESSED THE FIX, but lordy that was close to being *really* annoying.

(times)
  11:44  (mark this event: gonna learn to stream results)
  16:36  we suspect some version hell is breaking how our
         queries got parsed