refactor(STM): Refactor before add buttons (v)
Discussion: We are about to add new screens, and a big part
of that is adding new buttons (and taking buttons away etc).

We're not especially interested in fine-tuning bespoke layouts,
on for every screen. Rather, we want a function (if necessary)
that lays out the buttons in a sane way whether it's zero buttons,
one or many.

A one-size fits all grid would have been nice, but (as we'll
discuss in the future), that's not ready for prime time yet.

The focus of this commit is that it's simply a squash of many
commits that had small refactorings to accomodate the upcoming
changes.

Refactorings:

- functions as first class objects for responsive layout modes
- DRY up on nav click handler function
- buttons become an array of buttons - not constructed inline

TEMP SNAP WORKED

tiny snapshot (v)